requirements.txt
requests>=2.31.0
fake-useragent>=1.4.0
pycryptodome>=3.20.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
